WIP: home page is static but image
details page is dynamic
